+++ src/meetingmanager/test/AddMeetingPage2.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package userinterface;
+package meetingmanager.test;
 
 import javax.swing.JFrame;
 import javax.swing.SwingUtilities;
