+++ src/meetingmanager/test/EmployeePage.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package userinterface;
+package meetingmanager.test;
 
 import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.SwingUtilities;
-import meetingmanager.entity.Employee;
+
 
 import meetingmanager.entity.Employee;
 /**
@@ -29,11 +29,11 @@ public class EmployeePage extends javax.swing.JPanel {
     }
     
     
-
+/*
     public EmployeePage(Employee employee) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
-
+*/
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
